server/python/requirements.txt
pandas
numpy
pandas_ta
smartapi-python
pyotp
python-dotenv
